# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ObjectCategory = typeof ObjectCategory[keyof typeof ObjectCategory];


export const ObjectCategory = {
  Electronics: 'Electronics',
  Accessories: 'Accessories',
  Documents: 'Documents',
  Clothing: 'Clothing',
  Household: 'Household',
  Other: 'Other',
} as const;

export interface Object {
  id: number;
  userId: string;
  name: string;
  category: ObjectCategory;
  /** @nullable */
  description: string | null;
  /** @nullable */
  referenceImage: string | null;
  createdAt: string;
  updatedAt: string;
  observationCount: number;
}

export interface ObjectInput {
  /** @minLength 1 */
  name: string;
  category: ObjectCategory;
  description?: string;
  referenceImage?: string;
}

export interface Annotation {
  id: number;
  observationId: number;
  objectId: number;
  objectName: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  x: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  y: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  width: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  height: number;
  /** @exclusiveMinimum 0 */
  imageWidth: number;
  /** @exclusiveMinimum 0 */
  imageHeight: number;
  /** @nullable */
  classId: number | null;
  /** @nullable */
  className: string | null;
  annotationFormat: string;
  /** @nullable */
  datasetId: string | null;
  /** @nullable */
  trainingSessionId: string | null;
  createdAt: string;
}

export interface Observation {
  id: number;
  userId: string;
  objectId: number;
  objectName: string;
  /** @nullable */
  image: string | null;
  timestamp: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  detectionConfidence: number | null;
  source: string;
  createdAt: string;
  annotations: Annotation[];
}

export interface AnnotationInput {
  objectId: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  x: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  y: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  width: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  height: number;
  /** @exclusiveMinimum 0 */
  imageWidth: number;
  /** @exclusiveMinimum 0 */
  imageHeight: number;
  classId?: number;
  className?: string;
  annotationFormat?: string;
  datasetId?: string;
  trainingSessionId?: string;
}

export interface ObservationInput {
  objectId: number;
  image: string;
  timestamp?: string;
  latitude?: number;
  longitude?: number;
  locationName?: string;
  source?: string;
  annotations?: AnnotationInput[];
}

export interface HomeSummary {
  objectCount: number;
  observationCount: number;
  recentObservations: Observation[];
}