# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: objects
    description: Personal objects
  - name: observations
    description: Saved visual memories
  - name: annotations
    description: Manual visual annotations
  - name: dashboard
    description: Home screen summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /objects:
    get:
      operationId: listObjects
      tags: [objects]
      summary: List personal objects
      responses:
        "200":
          description: Personal objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Object"
    post:
      operationId: createObject
      tags: [objects]
      summary: Create a personal object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectInput"
      responses:
        "201":
          description: Created object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Object"
  /objects/{id}:
    get:
      operationId: getObject
      tags: [objects]
      summary: Get a personal object
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Personal object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Object"
        "404":
          description: Object not found
  /observations:
    get:
      operationId: listObservations
      tags: [observations]
      summary: List saved observations
      responses:
        "200":
          description: Saved observations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Observation"
    post:
      operationId: createObservation
      tags: [observations]
      summary: Save a manual observation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObservationInput"
      responses:
        "201":
          description: Created observation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Observation"
  /home-summary:
    get:
      operationId: getHomeSummary
      tags: [dashboard]
      summary: Get home screen summary
      responses:
        "200":
          description: Home summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeSummary"
  /demo-data:
    post:
      operationId: loadDemoData
      tags: [objects]
      summary: Load optional demo objects
      responses:
        "201":
          description: Demo objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Object"
components:
  parameters:
    ObjectId:
      name: id
      in: path
      required: true
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ObjectCategory:
      type: string
      enum:
        - Electronics
        - Accessories
        - Documents
        - Clothing
        - Household
        - Other
    Object:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/ObjectCategory"
        description:
          type: ["string", "null"]
        referenceImage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        observationCount:
          type: number
      required:
        - id
        - userId
        - name
        - category
        - description
        - referenceImage
        - createdAt
        - updatedAt
        - observationCount
    ObjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          $ref: "#/components/schemas/ObjectCategory"
        description:
          type: string
        referenceImage:
          type: string
      required:
        - name
        - category
    Observation:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        objectId:
          type: number
        objectName:
          type: string
        image:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        locationName:
          type: ["string", "null"]
        detectionConfidence:
          type: ["number", "null"]
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/Annotation"
      required:
        - id
        - userId
        - objectId
        - objectName
        - image
        - timestamp
        - latitude
        - longitude
        - locationName
        - detectionConfidence
        - source
        - createdAt
        - annotations
    ObservationInput:
      type: object
      properties:
        objectId:
          type: number
        image:
          type: string
        timestamp:
          type: string
          format: date-time
        latitude:
          type: number
        longitude:
          type: number
        locationName:
          type: string
        source:
          type: string
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/AnnotationInput"
      required:
        - objectId
        - image
    Annotation:
      type: object
      properties:
        id:
          type: number
        observationId:
          type: number
        objectId:
          type: number
        objectName:
          type: string
        x:
          type: number
          minimum: 0
          maximum: 1
        y:
          type: number
          minimum: 0
          maximum: 1
        width:
          type: number
          exclusiveMinimum: 0
          maximum: 1
        height:
          type: number
          exclusiveMinimum: 0
          maximum: 1
        imageWidth:
          type: number
          exclusiveMinimum: 0
        imageHeight:
          type: number
          exclusiveMinimum: 0
        classId:
          type: ["number", "null"]
        className:
          type: ["string", "null"]
        annotationFormat:
          type: string
        datasetId:
          type: ["string", "null"]
        trainingSessionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - observationId
        - objectId
        - objectName
        - x
        - y
        - width
        - height
        - imageWidth
        - imageHeight
        - classId
        - className
        - annotationFormat
        - datasetId
        - trainingSessionId
        - createdAt
    AnnotationInput:
      type: object
      properties:
        objectId:
          type: number
        x:
          type: number
          minimum: 0
          maximum: 1
        y:
          type: number
          minimum: 0
          maximum: 1
        width:
          type: number
          exclusiveMinimum: 0
          maximum: 1
        height:
          type: number
          exclusiveMinimum: 0
          maximum: 1
        imageWidth:
          type: number
          exclusiveMinimum: 0
        imageHeight:
          type: number
          exclusiveMinimum: 0
        classId:
          type: number
        className:
          type: string
        annotationFormat:
          type: string
        datasetId:
          type: string
        trainingSessionId:
          type: string
      required:
        - objectId
        - x
        - y
        - width
        - height
        - imageWidth
        - imageHeight
    HomeSummary:
      type: object
      properties:
        objectCount:
          type: number
        observationCount:
          type: number
        recentObservations:
          type: array
          items:
            $ref: "#/components/schemas/Observation"
      required:
        - objectCount
        - observationCount
        - recentObservations