# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Annotation {
  id: number;
  observationId: number;
  objectId: number;
  objectName: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  x: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  y: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  width: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  height: number;
  /** @exclusiveMinimum 0 */
  imageWidth: number;
  /** @exclusiveMinimum 0 */
  imageHeight: number;
  /** @nullable */
  classId: number | null;
  /** @nullable */
  className: string | null;
  annotationFormat: string;
  /** @nullable */
  datasetId: string | null;
  /** @nullable */
  trainingSessionId: string | null;
  createdAt: Date;
}