# lib/api-zod/src/generated/types/annotationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnnotationInput {
  objectId: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  x: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  y: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  width: number;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  height: number;
  /** @exclusiveMinimum 0 */
  imageWidth: number;
  /** @exclusiveMinimum 0 */
  imageHeight: number;
  classId?: number;
  className?: string;
  annotationFormat?: string;
  datasetId?: string;
  trainingSessionId?: string;
}