# lib/api-zod/src/generated/types/observation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Annotation } from './annotation';

export interface Observation {
  id: number;
  userId: string;
  objectId: number;
  objectName: string;
  /** @nullable */
  image: string | null;
  timestamp: Date;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  detectionConfidence: number | null;
  source: string;
  createdAt: Date;
  annotations: Annotation[];
}