# lib/api-zod/src/generated/types/observationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationInput } from './annotationInput';

export interface ObservationInput {
  objectId: number;
  image: string;
  timestamp?: Date;
  latitude?: number;
  longitude?: number;
  locationName?: string;
  source?: string;
  annotations?: AnnotationInput[];
}